Tried to fix broken models
It didn't work
At the moment corstFlipProto, corstProto, tInt3Proto and tInt4proto do
not draw. I'm not sure what makes these special compared to the rest. I
tried re-exporting them out of maya with new short mesh names (as
opposed to pasted_pasted_pasted_etc) and with a red lambert shader (as
opposed to the green blinn the rest use) to see if it would make a
difference to lighting, but it doesn't show up regardless. I've kept the
old version of those four files in the content/models/TempRails
directory in case you feel like trying stuff. I can't sync until I get
permissions tomorrow, but I'm done for the night. Even if those few
models don't render, the system is in place for it to all work smoothly
once we do get the real model in since all the dimensions are consistent