Texture offsets - planning for the future
Press space while boosting for something amazing
Did this while I was on the train. This is using a technique where the
model's texture is being generated during update() by drawing onto a
rendertarget, and this rendertarget cast as a texture2D is applied to
the mesh as a texture during draw. Generating the texture during the
draw function breaks things since it requires resetting the
GraphicsDevice rendertarget