Money gibs and enemy model
Added a swarmer model, made it the model of the enemy

Money gibs! - currently whenever a projectile hits something, it
produces a money gib. These use what could technically be considered AI
steering to first fly in a random direction before homing in on your
ship. Once it collides with the ship, you get the money