Asteroids and Electro turrets
Press right shift to place an electro turret. These build up a charge
and shoot up to 10 enemies within their radius with electricity
All enemies now have an ElectroDamage class attached to them - this
manages what happens when they're hit by electric damage (freeze in
place for a short amount of time, particle effect, later on there will
be animation and more effects too). Fire damage will have a similar one
of these once those turrets are put in.

Soon the turret bases will have a 'turret wheel' that lets you pick the
turret you place instead of the dodgy test buttons in place at the
moment

ASTEROIDS
Three different types, bigger ones spawn smaller once when destroyed.
The specifics of their health and sizes and stuff might change, but
there's more than enough in place for you to add physics to them jarred.
If possible do all the physics stuff in the Asteroid parent class using
the mass variable which then becomes unique to each subclass (so bigger
asteroids are heavier and require more effort to push). I put in the
names of a couple functions which I'd imagine will come in handy but
ultimately it's up to you

I'm in the middle of devising an elaborate system to spawn a balanced
random asteroid field, at the moment it just creates one asteroid at the
middle of each ast node of a random type

Handy protip, weapon 3 is currently an infinite ammo gun. wep 1 will
possibly be switched back to inifinite ammo - the issue is it has to be
nerfed in a way that it doesn't become the dominant strategy over all
other weapons because of the fact it's infinite ammo. It might be better
to just make ammo drops plentiful in the asteroid field because I'm
hesitant to give an infinite ammo weapon for combat situations