OOBB Update 1
An OOBB will now rotate correctly no matter what shape it is. This will
allow for trapezoidal shapes to be used when checking the collisions of
the flame turret shot (will be able to shoot in a cone like shape)

Also, this allows for OOBB's to be more tightly wrapped around an object
as they no longer need to be perfect rectangles or squares.