Shift code to WaveManager (along with references)
Moved the code in EnemyManager that spawned waves into a new WaveManager
class. Currently working in this class to create dynamic waves with
increasing difficulty. Will require for enemies to be able to damage
core before dynamic difficulty can be added.

Also a few tweaks to OOBB to make them a little more efficient,
calculating Normals is hard, a few experiments occurred, but no positive
results were found and all traces of experiment was removed.