New wave scheme as discussed in meeting
An attack now determines how many enemies there will be in total during
the attack

A new wave will not wait for enemies to be cleared, but will start 15
seconds after previous wave is finished spawnning.

Currently a placeholder wave creator is in place as the maths for
working out which enemy type and how many should be in a wave is hard
and I have to wake up first.