Gunner and Sneaker Paths
Gunner will no longer crash the game when a turret is placed while it is
in the map, I've actually seen them double back and go elsewhere to
avoid turrets I just placed, they work perfectly.

Sneakers will do the same at the moment, still yet to put in the player
damage thingy, but will not need to affect the enemy classes to do so, I
am now done playing with the enemies.