New wave scheme! (again)
An attack is made up of multiple waves, so far a level should have
around 8 attacks (can have more). Currently a wave only consists of one
enemy of that type, but that will change soon.

Attacks will calculate how many waves of each enemy type to spawn based
on the following calculations:

Swarmer - (int)((An + D) / 2) + 1
Gunner - (int)((An + D) / 3)
Sneaker - (int)((An + D) / 4)
Heavy - (int)((An + D) / 5)

Gunners do not appear until attack 2
Sneakers do not appear until attack 4
Heavies do not appear until attack 5

Amount of enemies in a given wave will be determined by the attack
number, the wave number in that attack and the type of enemy along with
a dynamic modifier.
This will be in the next update.