Enemy collisions, Swarmers and the drill
Enemies now store a list of circlecolliders, ultimately this will be
where all of their collision stuff is stored (luckily the enemy designs
are very round so we can get away with not using OOBBs

The swarmer has had some changes made so it reacts to getting hit by the
drill - it now has a state enum which includes every possible
state/condition it could be in. At the moment it only uses goToBase
(which calls a function where the stuff that was originally in update
has been dumped for the sake of keeping update clean) and hitByDrill,
but ultimately all will be used. Other enemies will use a similar system
but have their own state enum specific to their abilities