Asteroid Physics Update 2
They now bounce off the player ship, each other and walls. Still need to
update to bounce off derelict ships and sentries and when being shot.

Odd error occuring when an asteroid starts to move (sometimes) it will
think it's node position is some huge number and thus cause errors, I
have rectified this by simply deleting an asteroid when it thinks this,
but we need to find the source of the problem.