SMOOTH DIAGONAL WALL COLLISIONS!!
Fuck yeh, normals are my bitch!

However, using a circle collider to detect wall collisions has shown a
flaw in my circle collider code that must be fixed. The result of this
bug is that on concave wall joins the player can slip through the gap
between OOBBs.

I am currently working on a fix for this and hope to have one soon.