Manually entered map normal data
Dynamic normal detection just straight up will not work with how we have
the map set up, after over a week of trying to get it to be 100%
accurate I have simply entered the map collision normal into each OOBB
in a mapnode.

Still has an issue, it would have to be with how a ship deals with the
normals. Investigating now.