New Collision System - WIP
Bouncy as fuck, this is not intended to be a final solution. This is
just a build where it is IMPOSSIBLE to exit the map (without fucking
about with asteroids close to walls at least).

Just wanted to make sure if someone synced my changes they at least get
something that doesn't break the game.