Text fields for naming levels
To run the level creator, in game1 comment out initializeGame() and
runState = RunState.MainGame and uncomment initializeLevelCreator() and
runState = RunState.LevelCreator. the stray plus button that doesn't
react on mouse over will save the file when you click it. You have to
scoop it out of the bin/content folder and drag it into your VS project
to actually run it