Train Update - Placeholder Enemy Scripting
Extending enemy scripting beyond just simply which pathfinding AI to
use. A series of bools that define which behaviours each enemy class do
and do not partake in are now also read from the xml file.

The bools are implimented in the EnemyBehaviour class and are set to
some hardcoded default values. Hopefully either frank or myself can fill
out the load functions to actually set these values to their xml
courterparts. Also put those values in the xml file.

These values are, in their current state, ready to be used in the
decision tree for the enemy states when Jason begins work on that.