6/7 of the new SFX implemented
2 out of the 6 sounds need to be 3D sounds (laserHit and bombExplode)
however the apply3D() function is throwing an error so they are regular
sounds for now. Details for that are in the audioManager in case you
want to take a look and check to see if it works for you.

The 7th sound effect zap is also going to be a 3D sound however since
Jase said that there are lots of electricity particles being used it
would sound like a mess if they're not 3D. Once 3D is fixed, will
probably get Jase to implement all the zap sounds.

Extra effects: Pitch on the drill when colliding with objects is based
on whether the ship is boosting and the money gib sound increases in
pitch each time you collect a gib until the counter is reset after not
collecting gibs for a short time. Gib sound not implemented for ammo
drops (intended).

Also renamed the shockwave sound to missile because it was intended to
be for the infinite laser gun thing, not the shockwave.